Implementación de todos los loaders

+++ src/Classes/Protocols/Loadable.swift
@@ -11,7 +11,7 @@ import SDOSSwiftExtension
 public protocol Loadable: UIView {
     static func createLoader(loaderType: LoaderType, inView view: UIView, size: CGSize?) -> LoaderObject
     func show(loaderObject: LoaderObject, delay: TimeInterval)
-    func setProgress(loaderObject: LoaderObject, value: Double)
+    func setProgress(loaderObject: LoaderObject, value: Float)
     func setText(loaderObject: LoaderObject, text: String)
     func hide(loaderObject: LoaderObject)
